A7

+++ todo/app/Models/MitgliederAufgabenModel.php
@@ -14,7 +14,14 @@ class MitgliederAufgabenModel extends Model{
 
     protected $useTimestamps = false;
 
+    /**
+     * speichert/updated einen Batch an AufgabenID und MitgliederId Tupel
+     * @param $aufgabenid
+     * @param $mitgliederids
+     * @throws \ReflectionException
+     */
     public function upsert($aufgabenid, $mitgliederids){
+        if($mitgliederids == []){return false;}
         $this->where('aufgabenid', $aufgabenid)->delete();
 
         $data = [];
